soccer-prediction-advanced-models/requirements.txt
numpy
pandas
scikit-learn
xgboost
lightgbm
tensorflow
matplotlib
seaborn
pyyaml
jupyter
notebook